Implement capture intent management and enhance export activity handling

+++ apps/desktop/src-tauri/src/recording/mod.rs
@@ -336,7 +336,7 @@ fn resolve_region_target(rect: RegionRect) -> Result<CaptureTarget> {
             // The frontend Region Picker passes coordinates in PHYSICAL pixels,
             // but xcap's Monitor bounds are in LOGICAL pixels.
             // We must un-scale the physical center point to find the matching monitor.
-            let scale = display_scale_factor(monitor) as f32;
+            let scale = display_scale_factor(monitor);
             let cx = (center_x as f32 / scale).round() as i32;
             let cy = (center_y as f32 / scale).round() as i32;
 
@@ -349,7 +349,7 @@ fn resolve_region_target(rect: RegionRect) -> Result<CaptureTarget> {
         .context("unable to locate the display containing the selected region")?;
 
     let scale = display_scale_factor(&monitor);
-    let scale_f32 = scale as f32;
+    let scale_f32 = scale;
 
     // Convert the physical rect back to logical pixels so it shares the same
     // coordinate space as the monitor for clamping.
@@ -447,25 +447,25 @@ pub struct RecordingOptions {
     #[serde(default)]
     pub microphone: bool,
     /// Microphone device ID (None = default device).
-    #[serde(default)]
+    #[serde(default, skip_serializing_if = "Option::is_none")]
     pub microphone_device_id: Option<String>,
     /// Capture camera video.
     #[serde(default)]
     pub camera: bool,
     /// Camera device ID / DirectShow device name (None = first available).
-    #[serde(default)]
+    #[serde(default, skip_serializing_if = "Option::is_none")]
     pub camera_device_id: Option<String>,
     /// Capture frame rate. `None` (or out of the supported 24..=240 range)
     /// falls back to [`RECORDING_FPS`]. The pacer and encoder both run at this
     /// rate; values above the monitor's refresh just duplicate frames, so the
     /// UI gates the offered options by the detected display refresh.
-    #[serde(default)]
+    #[serde(default, skip_serializing_if = "Option::is_none")]
     pub fps: Option<u32>,
     /// Capture quality tier — `"auto"` (default), `"balanced"`, `"high"`, or
     /// `"pristine"`. `"auto"`/unknown resolve against the detected encoder
     /// (hardware → high, software → balanced). See
     /// [`crate::encoder::RecordingQuality::resolve`].
-    #[serde(default)]
+    #[serde(default, skip_serializing_if = "Option::is_none")]
     pub quality: Option<String>,
 }
